Separate blockmaker for each blockchain (bitcoin will follow)
