+++ src/bitcoin/BitcoinUtils.cc
@@ -99,20 +99,6 @@ int64_t GetBlockReward(int nHeight, const Consensus::Params& consensusParams)
 #endif
 
 
-uint64_t GetBlockRewardBytom(uint64_t nHeight)
-{
-	//	based on bytom's mining.go (BlockSubsidy function)
-	const uint64_t initialBlockSubsidy = 140700041250000000UL;
-	const uint64_t baseSubsidy = 41250000000UL;
-	const uint64_t subsidyReductionInterval = 840000UL;
-	if(nHeight == 0)
-	{
-		return initialBlockSubsidy;
-	}
-	return baseSubsidy >> (nHeight/subsidyReductionInterval);
-}
-
-
 #ifdef CHAIN_TYPE_SBTC
 
 CTxDestination DecodeDestination(const std::string& str) {
