+++ src/bytom/BytomUtils.h
@@ -1,0 +1,8 @@
+#ifndef BYTOM_UTILS_H_
+#define BYTOM_UTILS_H_
+
+#include "CommonBytom.h"
+
+uint64_t GetBlockRewardBytom(uint64_t nHeight);
+
+#endif //  BYTOM_UTILS_H_
